Add support for type conversion for Integer Datatypes
Integer datatypes can now be mixed in operations
Values smaler than i32 will be converted to i32
Values bigger than i32 remain intact.
In binary operations, values are converted to the bigger type.
When assigning, values are truncated or expanded to fit the target size
Added a correctness test for the edge cases
Added a test case for multiple expressions that should cause the
types to change.

Co-authored-by: Mathias Rieder <mathias.rieder@gmail.com>