Added Float and Int casts
Added tests to cover the float to int cast and vice  versa

- Conversion to LREAL is still missing, (LREAL support in general)
Test : int_bigger_than_float_converted_to_double is ignored until
double/ LREAL support is implemented

- No Correctness tests were added

Cleanup in Codegen

Refactored the promote if needed method to call a get_bigger_type method
instead of repeating the work.
The current implementation requires cloning of all types, we should
still look into this.