struct initializatino falls back to defaults
if the initial value of a struct-member is not provided
when initializing an instance like i : MyStruct := (a:=1, b:=2...)
then it will fallback to the member's initialization defined in the
struct's declaration or the datatype's default.

The index now contains the list of member-names for a container
and the inital values of member-variables as well.