arena implementation for const-expressions (#284)
arena implementation for const-expressions

the index now offers a arena for const rexpressions so they
can be later evaluated and replaced (with the same id)
so validation and codegen can use numbers that are
available at compile-time

const expressions include:
- intializers    	x : INT := LEN - 1
                                   -------
- array-dimensions   	x : ARRAY[0..LEN-1] OF INT;
                                  -  -----
- String-length		x : STRING(LEN-1);
			           -----

introduce TypeSize enum to adress compile-time literals

the TypeSize enum can either represent a const-expression that
is meant to be evaluated at compile time (e.g. LEN - 1) or a
compile time literal (e.g. the default length of a String).