Support for constants issue208 (#279)
* arena implementation for const-expressions

const expressions are no longer stored directly in the index
but they are stored in an arena. This way it is easier to
replace them later when we want to resolve the expressions
into literals.

* const expr are now managed in the const-arena

the index now offers a arena for const rexpressions so they
can be later evaluated and replaced (with the same id)
so validation and codegen can use numbers that are
available at compile-time

const expressions include:
- intializers    	x : INT := LEN - 1
                                   -------
- array-dimensions   	x : ARRAY[0..LEN-1] OF INT;
                                  -  -----
- String-length		x : STRING(LEN-1);
			           -----

* introduce TypeSize enum to adress compile-time literals

the TypeSize enum can either represent a const-expression that
is meant to be evaluated at compile time (e.g. LEN - 1) or a
compile time literal (e.g. the default length of a String).

* easy compile-time resolving of constants

literals and refrences to ints and bools

* eval constants (binary expressions)

* introduce !arith macro to avoid code duplication

* support bitwise operators in const intializers

special handling for unsigned numbers!

* introduce const literals to codegen

* improved codegen and tests

* improved generate_global_constants

* refactor dimensions so const-expr can be used

Dimensions keep the ast-statements as start and end which
eventually will be replaced by a LiteralInteger if it can be
evaluated.

this allows to collect types and constants in the visit-phase
and following that, replace the dimensions that consist of
complex expressoins (e.g. "a + b") with a literal-integer
once all constants are known.

* ntermed index-step: try_resolve_pending_const_expressions

* introduce arena changes and fix tests

* added qualified const references

qualified references for now can be of the form:

POU.variable

referencing fields of const structs (E.g. a.b.c) are
not supported yet

* cleanup and documentation

tests that shows that constants can be used in body
the current implementation creates global variables for constants
and references these variables in expressions.

global's could be inlined to create more efficient code
see issue #291