use correct type when generating array literals
asking the DataTypeInformation for the type's name (to then
fetch the llvm-type registered for that name) is a problem for
STRINGs, because the DataTypeInformation always returns "STRING"
as name, which maps to the llvm type [81 x i8] and therefore
ignores the literal's real type.

so we need to use the DataType's name, not the DataTypeInformation's
name.

fixes #417