correctly annotate adresse-of expressions
addressof-expressions may create a new pointer-type that we did not see
before, so we the resolver may create this type on the fly.

These new types need to be visible for the resolver while resolving
a body. This requires some particular order of operations to avoid
life-time issues of the 'index' and the 'new_index'.