annotate and generate STRING-literals precisly according to their length (#421)
he annotator tells us exactly how long every string-literal is by craeting dedicated STRING-types for them. The expression generator now applies two different strategies:

when generating string-literals in declarations (outside of a body) we generate the full length of the string-literal. So a STRING[80] = 'a' will generate 80 additional \00 to fill the full vector
when generating string-literals in the body we generate only the string's real length. So an assignment: x := 'a'; will generate an vec of length 2 with 'a' and \00 in it (so we don't fill x.
fixes #417