turn bool into i8
generating bool as i1 makes it impossible to have bool-arrays
because an array-access (x[2]) can only adress (read/write) a
whole byte, not a single bit.

bit-access relies on bool beeing a single bit so I introduced an
internal (hidden) type called '__U1' that is really one bit.
If we detect a bool in bit-access, we internally switch form
BOOL to __U1 to get the original behavior

fixes #420