Issue 480 add POUs to index (#487)
added PouIndexEntry
the pouIndexEntry offers additional information of pou's

program instance variables are no longer stored in the index's
global_variables map. the VariableIndexEntry is now stored in the
Program's POU index entry. We do this because:

there is now a strong connection between the pou and the instance
variable, where previously it was a lose connection via the pou-name
methods and actions are indexed as POUs as well.

use the new pou-information to make some linker-decisions in the
resolver more concise and clearer code

this way we loosen the dependency between the pou's name and
the implementation's name. So far the pou's implementation needs
to have the exact same name like the POU.

linkage and generic information is no longer stored with the pou's
struct type. this information always felt a bit alien for the
struct-type.