Issue 468 function calls should pass single args instead of pou like instance struct (#492)
introduce single parameter passing for functions

passing parameters to functions is treated differently from passing parameters to other POUs like Programs or FunctionBlocks. While parameters for calls to Programs and FBs are optional, the also hold a state in between two calls (so call#2 can see the parameters passed to call #1 if the call#2 does not re-set the same parameters).

While Programs and FBs take a single parameter (a struct with all fields), functions take every single field as a dedicated argument. This means that calls to functions work very differently from calls to Programs/FBs

e.g. call to a program PRG

1. find instance-struct for PRG -> e.g. g_prg
2. fill all fields passed to this call (e.g. PRG(a:=1, b:=2))
    g_prg.a = 1;
    g_prg.b = 2;
3. call the program
   prg(g_prg)

e.g. call to a function Foo
1. call the function and pass all parameters
   foo(1,2)