improved validation of FUNCs without return types
Functions with missing return types used to report a problem on the
whole POU (declaration and body). I added an additional location field
to the POU (AST) that stores the location of the POU's name so we can
report the error more nicely where it actually happened.

also activated dead test (pou_validation_tests)