constant propagation in ranged type limits
added a test that shows that constants are propagated even in
ranged-type's limits, so even if you use (const-)expressions (a+b)
to define the limits of a ranged type, the const-expressions
will be evaluated during compile time and the result will be used.
- as in contrast we would do this calculation all the time at
runtime.

the behavior was introduced in PR #504 (016328)

fixes #350