Add two tests related to VAR_OUTPUT
These tests currently fail due to VAR_OUTPUT getting lowered into a
reference.

One tests behavior when an output is used inside the function-block, but
not assigned in the call.  The other tests whether outputs are persisted
across invocations.

For now, both tests are skipped as they are failing due to issue #530.