avoid re-registering generic implementations (#641)
when calling a generic function, we register an implementation and
a POU in the index to later link against it. If multiple calls discover
the same generic implementation (e.g. INT-implementation), they were
added mulitple times to the index and the fix implemented in #354 caused
a problem.

ImplementationIndexEntry move back to IndexMap - we only keep one entry
per name - there is no value since the POUIndexEntries handle the
duplication validation.

When importing POUIndexEntries we ignore automatically generated
entries if they are already in the index.

fixes #637