use by-ref out variable for aggreate return types
Functions that return an aggregate type (string, array, struct) no
longer return the value as the function's return type, they rather
return the value via an out-pointer that needs to be passed by the
caller.
old: `myArr := foo(1, 2, 3);`
will be generated automatically as: `foo(&myArr, 1, 2, 3);`

The out parameter is inserted at position 0 into the parameter list.

To properly handle these situations this commit introduces the
ExpressionValue in expression_generator.rs to easily distinguish
lvalues (something we can get a pointer to) from rvalues. Previuosly
the expression_generator was always returning rvalues (it automatically
loaded the pointers) which makes it unenecessary hard to use memcpy
instead of stores when applicable.

This commit also fixes wrong behavior for casted-string-literals for the
situations: STRING#"abc" and WSTRING#'abc'.

fixes #652, #446, #418