Fix casting for function arguments
Our generated IR signature was previously wrong when running with the verify feature.
This was mainly due to us either not casting literals to their hinted type or simply
not bitcasting non-aggreggate arguments at all. This commit fixes this issue and
should now generate the correct IR, such that the verify feature no longer fails
for these kind of cases.