Allow arrays to be passed as function arguments (#764)
Trying to pass an array as a function argument triggered a validation error
because we we're checking if the passed argument is of the same type as the declared
parameter. The detection of such code in the is_same_type_class function was non-existing however.
This commit adds such a check such that passing array arguments is allowed.