Annotate reference as POU if we're dealing with a call statement (#919)
Fixes an issue where calling a function with the same name as a locally defined variable would incorrectly annotate the call as a variable. For example calling an external function `foo` inside a function `main` with a locally defined variable `foo` would incorrectly annotate the function call `foo` with `main.foo`.