format: Convert tabs to spaces
Project mostly use spaces over tabs. When mixing tabs and spaces this usually
makes formatting issues and also when changing those in commits it will make lot
of git noise. We will force spaces so just convert everything to tabs already.

We did also have some users who tab wide was 2 and some 4. This is very common
in projects to happen. Using pre-commit for these we can basically prevent this.

Commit was generated with:

    pre-commit run --all-files remove-tabs