refactor: add Pre-commit + editorconfig + gitattributes (#1034)
* Fix indentation levels cause by tabs

In many places indentation level is little of because tabs where used.
Probably someone might have example tabs size 2. Let's fix these by hand
as automatic tools cannot fix these for us. They can prevent them from
happening but fixing is not easy.

* Fix unicode replacement chars

These seems like accidents so remove those so we can start verify whole
repo against those. In combined.st I changed language to english as it
seems it should be.

* Change all line-endings to LF in whole project

There where not many files which where with CRLF. So it looks like we
should always use LF. Let's fix couple files and then make sure this
does not happen again. I also have some other problems with some of
these files as I get strange douple newlines.

* Add .gitattributes to force line-ending LF

Force line-endings to LF. It could probably also be auto, but unless
someone complain let's keep it LF. This will help that we have unified
way for line-endings so we do not get those accidental commits where
whole file is just big diff because line-ending change.

* Add .pre-commit-config.yaml and .editorconfig

With these we can automate lot of formatting issues with all files. We
also get cleaner commits as every commit will be check with various
small checkers.

* Add editorconfig extension to devcontainer

Now that we have editorconfig file it make sense to add it also to
VScode as it does not support it out of the box.

* ci: Remove submodules checks as we do not have any

We do not have submodules so we do not need to check those.

* format: Strip trailing whitespaces

We want to get rid of trailing whitespaces completly as they make just git
noice. Much better to start using automated tools to get rid of them once and
not getting them back again. This way git history will be cleaner and review
easyer.

Commit was generated with:

    pre-commit run --all-files trailing-whitespace

* format: File should have exacly one new line end of them

It is good practice that every file has one new line. It is not now days so
mandatory but it also is not nice if file has lot of newlines end of it. We will
use pre-commit which takes automatically care about this so let's fix all.

Commit was generated with:

    pre-commit run --all-files end-of-file-fixer

* format: Run Black for whole repo

Currently we only have one python file but is is nice if we have same formatting
for all files. We will use Black for this.

Commit was generated with:

    pre-commit run --all-files black

* format: Convert tabs to spaces

Project mostly use spaces over tabs. When mixing tabs and spaces this usually
makes formatting issues and also when changing those in commits it will make lot
of git noise. We will force spaces so just convert everything to tabs already.

We did also have some users who tab wide was 2 and some 4. This is very common
in projects to happen. Using pre-commit for these we can basically prevent this.

Commit was generated with:

    pre-commit run --all-files remove-tabs

* Regenerate test snapshots

As we have changed style and example removed whitespaces we need to regenerate
all test snapshots.

Commit was generated with:

    ./pre-py/snap-pre.sh; cargo test --all; ./pre-py/snap-rework.sh; cargo test --all; ./pre-py/snap-rework.sh

---------

Co-authored-by: Kari Argillander <kari.argillander@fidelix.com>