chore: Update Cargo.lock for security and refactor deprecated chrono functions (#1045)
* fix: update Cargo.lock for security updates

Rustix needed an update, ran cargo update to address the issue

* Use timestamp_nanos_opt() over timestamp_nanos() in tests

timestamp_nanos() is deprecated and we can replace it very easily in
unittests with timestamp_nanos_opt().unwrap(). These are unittests and
we do not except timestamp_nanos_opt() to fail so unwrap is totally ok.

* Use timestamp_nanos_opt() over timestamp_nanos()

timestamp_nanos() is deprecated and we should replace it with
timestamp_nanos_opt(). Before it would just panic if range was not in
between 1677-09-21T00:12:44.0 and 2262-04-11T23:47:16.854775804. Now we
can decide what to do in which situation.

I tried to follow what ever error stategy function did before.

* Replace deprecated from_utc() with from_utc_datetime()

Function from_utc is deprecated. I choose to use
TimeZone::from_utc_datetime which is pretty similar.

---------

Co-authored-by: Ghaith Hachem <ghaith.hachem@gmail.com>
Co-authored-by: Kari Argillander <kari.argillander@fidelix.com>