feat(#826): refactor the diagnostics to be more consistant
Diagnostics are now created using builder methods, and are no longer
enums. A severity field now indicates the default severity of a
diagnostic. This can be overridden in a diagnostian. This is yet to be
implemented.
We now use Anyhow to report some errors to simplify conversions between
our Diagnostic type and some rust types.