feat(validation): Add further checks for enum validation (#1064)
Previously the enum validation checked whether the datatypes of the left- and right-hand side are equal and if not returned an error. This version adds some further checks to see if an assignment would yield a value that satisfies all the enum's variant values. For example for TYPE Color : (red, green, blue := 5); END_TYPE the requirement would be that the right-hand side of an assignment evaluates to a literal integer of 0, 1 or 5. With that in mind, we can now give more precise error/warning messages.