feat: registery based diagnostician (#1077)
* feat: add a Dagnostics registry that controls severity

An Diagnostic registry is now responsible for adding a severity instead
of the diagnostic itself. The registry is the default diagnostics
assessor. Diagnostics no longer have an own severity field as a result.

* Rename error to new
* Diagnostics config option
* feat: add registry serializer

Added the option to load a diagnostics configuration
Added tests for the configuration
Added the basic api to print the configuration

* feat: add command line option to parse plc_diagnostics

A  new parameter (`--error-config`) can now be used to include a json
file with the default behaviour for errors

* feat: add an explain flag to the compiler

Error messages now show the error code in the report
At the end of the compilation a hint is shown to use the new explain
flag for more info

* feat: add the possiblity to print the error config

The error configuration can now be printe using `plc config diagnostics`

Added documentation about the new commands to the book
* Make sure windows runs on 1.75
* Reword some errors