fix: Derive root directory from unit (#1153)
Fixes a bug for Windows where compilation would panic if the compiler was executed in another drive than where a project was located in. For example if `plc build D:/project/conf/plc.json [...]` was executed from the `C:/` drive, the pipeline would try to strip the root `C:/` from the path `D:/project/...` which panics because of the method [`strip_prefix`](https://doc.rust-lang.org/std/path/struct.Path.html#method.strip_prefix). For linux and macOS this was no problem because the root is always `/`.