feat: Be more lenient when dealing with integer values in conditions (#1186)
Resolves https://github.com/PLC-lang/rusty/issues/1148

Previously, `if` and `while` statements expected the condition to be of type boolean and otherwise returned an error. Some projects make a lot of use of integer values as conditions, hence this PR distinguish between these two cases and returns a warning for the latter. Also values that evaluate to `0` or `1` are allowed without any warning / error.