fix: Array sizes in DWARF (#1190)
This PR fixes a bug in which we would generate incorrect array sizes due to an off by one issue. This is because for an array such as `ARRAY[1..5]` we pass the range `1..5` to an inkwell function which calculates the size by subtracting the start from the end i.e. `5 - 1`. ST array subranges are inclusive however, hence adding 1 to end value should fix this.