refactor(AST): introduce AstVisitor trait (#1231)
* refactor(AST): introduce AstVisitor trait

The AST-Visitor trait allows generic visiting of AST-nodes. A default
Walking behavior is implemented for each AstStatement but it can be
altered by any implementor. When overriding a visit_XXX method, the
implementation can decide to continue with the default walking behavior
(by calling the walk function on the passed AstStatement-Struct, to skip it,
or to continue with an alternative walking bahavior.

removed unused AST element CastStatement