test: add alias to run `cargo t --workspace` on a single thread (#1337)
Sometimes running cargo t --workspace can result in a lock-up of the machine it is being run on, which is a big problem when working remotely via ssh, since the only way to get your machine back to a working state is a soft-reset. This commit provides a convenient alias to run workspace tests on a single thread.