lowering: Lower property references to get or set calls
This commit transforms references of properties such as `myFb.myProp := 5` and `myVar := myFb.myProp`
into setter and getter methods respectively. Broadly speaking when dealing with assignments
the lowerer will check if the left side is a property and transform the whole node into a set call
by `<lhs_node>(<rhs node);`. Otherwise, the visitor will visit every statement and check if the
reference is a property and if so replace it with a getter call.