feat: Properties (#1396)
This PR introduces the `PROPERTY` language feature. Properties in itself are getter and setter methods defined in a (stateful) POU and have the form of
```
FUNCTION_BLOCK fb
  VAR
    localPrivateVariable : DINT;
  END_VAR

  PROPERTY myProp : DINT
    GET
      // other statements potentially modifying the `localPrivateVariable`
      myProp := localPrivateVariable;
    END_GET
    
    SET
      // other statements potentially modifying internal variables
      localPrivateVariable := myProp;
    END_SET
  END_PROPERTY
END_FUNCTION_BLOCK
```
They are invoked by other POUs by simply accessing the property name. For example
```
FUNCTION main
  VAR
    x : DINT
    instance : fb;
  END_VAR

  instance.myProp := 5; // Internally resolves to `instance.__set_myprop(5)`
  x := instance.myProp; // Internally resolved to `x := instance.__get_myprop(5)`
END_FUNCTION
```
Specifically, references to properties always resolve to get calls with the exception in assignments when the property itself is accessed on the left-hand side.

The compiler will initially handle these properties as dedicated `Property` nodes in the `CompilationUnit` to have a 1:1 representation of the user code and only later on lower these properties to methods and property references to calls.

---------

Co-authored-by: abroooo <61380142+abroooo@users.noreply.github.com>