feat: Global Namespace Operator (#1442)
Introduces the global namespace operator to access global variables directly when they would otherwise be shadowed by a local variable. The operator is a simple dot that is prefixed before the variable name, for example `.foo` would access the global variable `foo`.

This feature is achieved by introducing a `ReferenceAccess::Global(AstNode)` variant and a new resolving strategy to annotate such nodes accordingly.