feat(cli): Introduce `--ast-lowered` flag
Changes the behavior of the `--ast` flag to always emit the pre-lowered AST while introducing a `--ast-lowered` that matches the current `--ast` output.

Still not perfect in that one can not adjust at what stage to emit the AST (i.e. similar to the participant trait) but good enough for now to peek at the AST before ANY lowering step was executed.