feat: support `SUPER` keyword (#1445)
Adds support for the SUPER keyword and usage.
Rather than adding a pointer to each derived POU, SUPER is only represented in the AST - references to it use the already existing __<parent> member variable, either directly (when SUPER^ is dereferenced) or its address via REF(__<parent>) (when it is not). This lowering is done in its own submodule in the Inheritance-lowerer before lowering the inheritance chains.