feat: `FB_INIT` user code initialization (#1458)
Adds initialization logic for user-defined `FB_INIT` methods and updates the user-documentation.
Struct-types now get an additional init-function for user-initialization, which will call the `FB_INIT` if present and then initialize each member, similar to the already present `__init_<type>` functions. The user_init functions (`__user_init_<type>`) are guaranteed to be called after the implicit init is finished to allow pointers to be initialized beforehand.