fix: Resolve nested assignments
Something like `var : StructA := (instanceB := (instanceC := (value :=
5)))` would previously not work because the resolver had incorrect
metadata when recursively descending into these assignments.
Specifically it always assumed the name of the left-hand side to be
whatever the first element was when visiting the right-hand side as a
whole, i.e. it assumed `StructA` when in reality it needed to be
`StructA` for `(instanceB := (...))`, `StructB` for `instanceC := (...)`
and `StructC` for `value := 5`.