refactor: Flag POINTER TO keyword as type-unsafe
Any variable declaration making use of the `POINTER TO` keyword will not be flagged as type-unsafe in the `DataType::PointerType` variant. With that change the validation can make use of that flag and ignore type-checks when dealing with ADR calls which by definition return an LWORD, e.g. `typeUnsafePtr := ADR(someVariable);`.

This commit reflects these changes from the parser up to the resolver. Validation changes have yet to happen.