fix: Disable type-checks for `POINTER TO` variables (#1482)
This PR makes `POINTER TO` declared variables type-unsafe, such that no type-checking is done. For example, while incorrect, no validations _should_ trigger for the following code
```
VAR
    stringVar : STRING;
    unsafePtrA : POINTER TO DINT := ADR(stringVar);
    unsafePtrB : POINTER TO DINT := REF(stringVar);
END_VAR
```

The reason for such behavior mostly stems from the fact that CodeSys behaves in that way and we want to
be compatible with CodeSys to some extent.