fix: Handle generics with no return type (#1484)
The following code example will panic on master because we except a
return type to be present. However, with https://github.com/PLC-lang/rusty/pull/1103
this is no longer the case. This commit fixes the issue by assigning
a `VOID` datatype if the index was not able to find a return type for
generics.

```
FUNCTION genericVoidFunction<U: ANY>
  VAR_INPUT
    in : U;
  END_VAR
END_FUNCTION

FUNCTION main
  VAR
    localVar : DINT;
  END_VAR

  genericVoidFunction(localVar);
END_FUNCTION
```